react intit enhancement
