+++ local-movie-streamer-backend/.env
@@ -1,2 +1,2 @@
-IP_ADDRESS="192.168.1.100"
-
+# IP_ADDRESS="192.168.1.100"
+IP_ADDRESS="172.23.144.1"
