+++ local-movie-streamer-frontend/.env
@@ -1,1 +1,2 @@
-VITE_IP_ADDRESS="192.168.1.100"
+# VITE_IP_ADDRESS="192.168.1.100"
+VITE_IP_ADDRESS="172.23.144.1"
